+++ common/message/codec_header.hpp
@@ -30,7 +30,7 @@ namespace msg
 class CodecHeader : public BaseMessage
 {
 public:
-    CodecHeader(const std::string& codecName = "", size_t size = 0)
+    CodecHeader(const std::string& codecName = "", uint32_t size = 0)
         : BaseMessage(message_type::kCodecHeader), payloadSize(size), payload(nullptr), codec(codecName)
     {
         if (size > 0)
